pyproject.toml
[project]
name = "mail-agent"
version = "0.1.0"
description = "Mail Agent for processing Gmail with user-defined rules"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.5.2",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "click>=8.1.3",
    "python-dotenv>=1.1.0",
]

[project.scripts]
mactl = "mail_agent.cli:cli"
